refactor: ListingProcess back to ListingView
