+++ sites/public/src/ListingView.tsx
@@ -3,54 +3,125 @@ import ReactDOMServer from "react-dom/server"
 import Markdown from "markdown-to-jsx"
 import moment from "moment"
 import {
-  Listing,
-  ListingEvent,
-  ListingEventType,
-  ListingApplicationAddressType,
   ApplicationMethod,
   ApplicationMethodType,
+  Listing,
+  ListingApplicationAddressType,
+  ListingEvent,
+  ListingEventType,
 } from "@bloom-housing/backend-core/types"
 import {
   AdditionalFees,
   Description,
-  GroupedTable,
-  ImageCard,
+  FavoriteButton,
   GetApplication,
+  GroupedTable,
+  Heading,
+  ImageCard,
+  InfoCard,
+  LeasingAgent,
   ListingDetailItem,
   ListingDetails,
   ListingMap,
-  OneLineAddress,
-  SubmitApplication,
+  ListingUpdated,
   ListSection,
+  OneLineAddress,
+  OpenHouseEvent,
+  ReferralApplication,
   StandardTable,
+  SubmitApplication,
   t,
-  FavoriteButton,
-  InfoCard,
-  Heading,
+  Waitlist,
 } from "@bloom-housing/ui-components"
 import {
   cloudinaryPdfFromId,
   imageUrlFromListing,
-  occupancyTable,
   listingFeatures,
+  occupancyTable,
 } from "@bloom-housing/shared-helpers"
 import dayjs from "dayjs"
 import { ErrorPage } from "../pages/_error"
 import {
   getGenericAddress,
   getHmiSummary,
+  getImageCardTag,
+  getListingTag,
   getListingTags,
   getUnitGroupSummary,
   openInFuture,
-  getListingTag,
-  getImageCardTag,
 } from "../lib/helpers"
-import { ListingProcess } from "../lib/ListingProcess"
+
+interface ListingProcessProps {
+  listing: Listing
+  openHouseEvents: ListingEvent[]
+  applicationsClosed: boolean
+  hasNonReferralMethods: boolean
+  applySidebar: () => JSX.Element
+}
 
 interface ListingProps {
   listing: Listing
   preview?: boolean
   allowFavoriting?: boolean
+}
+
+export const ListingProcess = (props: ListingProcessProps) => {
+  const {
+    listing,
+    openHouseEvents,
+    applicationsClosed,
+    hasNonReferralMethods,
+    applySidebar,
+  } = props
+
+  // TODO: Need help understanding why isWaitlistOpen, name and website are unhappy here
+
+  return (
+    <aside className="w-full static md:mr-8 md:ml-2 md:border border-gray-400 bg-white">
+      <ListingUpdated listingUpdated={listing.updatedAt} />
+      {openHouseEvents && <OpenHouseEvent events={openHouseEvents} />}
+      {!applicationsClosed && (
+        <Waitlist
+          isWaitlistOpen={listing.isWaitlistOpen}
+          waitlistMaxSize={listing.waitlistMaxSize}
+          waitlistCurrentSize={listing.waitlistCurrentSize}
+          waitlistOpenSpots={listing.waitlistOpenSpots}
+        />
+      )}
+      {hasNonReferralMethods && !applicationsClosed && applySidebar()}
+      {listing?.referralApplication && (
+        <ReferralApplication
+          phoneNumber={
+            listing.referralApplication.phoneNumber ||
+            t("application.referralApplication.phoneNumber")
+          }
+          description={
+            listing.referralApplication.externalReference ||
+            t("application.referralApplication.instructions")
+          }
+          title={t("application.referralApplication.furtherInformation")}
+        />
+      )}
+      {openHouseEvents && (
+        <div className="mb-2 md:hidden">
+          <OpenHouseEvent events={openHouseEvents} />
+        </div>
+      )}
+      <LeasingAgent
+        listing={listing}
+        managementCompany={{
+          name: listing.managementCompany,
+          website: listing.managementWebsite,
+        }}
+      />
+      {listing.neighborhood && (
+        <section className="hidden md:block aside-block">
+          <h4 className="text-caps-underline">{t("listings.sections.neighborhoodTitle")}</h4>
+          <p>{listing.neighborhood}</p>
+        </section>
+      )}
+    </aside>
+  )
 }
 
 export const ListingView = (props: ListingProps) => {
